netsync: Make hdr sync transition concurrent safe.
The current approach to logging during the initial headers sync and
transitioning from the initial headers sync mode to the initial chain
sync mode relies on the fact the entire sequence is mutually exclusive,
as is currently the case, but it would result in logic races if parallel
header downloading is ever implemented during the headers sync process.

There are also no immediate plans to implement parallel header
downloading, in part because doing so would require an entirely new and
more capable wire message to be implemented first, but also because it's
already such a tiny fraction of the overall sync time and headers don't
grow at a very fast rate, so it will be a very long time before any
bottleneck there would become an issue.

However, it is worthwhile to address it now even though it's not
technically required yet since it might not be super obvious later.

Consequently, this reworks the logic related to transitioning to the
initial headers sync done state slightly to be independently concurrent
safe by adding a mutex to the header sync state and ensuring the flag
that tracks its completion, which is also separately an atomic for
performance, is re-checked and updated under that mutex to ensure the
transition is idempotent.

It also gates the header sync progress logging behind the initial header
sync mode flag since it is technically more accurate even though it does
not really matter so long as there is only a single header sync peer.

This is a part of the overall effort to convert the code related to
handling the various sync manager events to synchronous code that runs
in the various caller goroutines.