netsync: Simplify request maps.
This simplifies the request tracking for blocks, transactions, and mix
messages by making use of a single global map for each type that tracks
the peer instead of a global map plus an additional per-peer map for
each type.

Note that this change does make it very slightly more costly in terms of
CPU usage to determine the in-flight requests for a specific peer, but,
in practice, the new approach ends up being slightly faster overall
despite that due to the combination of less map management performed
under a mutex as well as the infrequent need to determine the in-flight
requests for a specific peer.

It also slightly reduces memory usage due to the removal of 3 maps per
connected peer which totals to between 24 and 375 maps with typical
default connection settings.