netsync: Rework external request from peer.
This reworks the logic that allows code external to the sync manager to
request blocks, votes, and treasury spend transactions to both simplify
and optimize it.  The method now makes use of the same logic the sync
manager itself uses when requesting data and no longer potentially
returns an error.

Of particular note is that it now uses the same much more efficient code
path that the sync manager itself uses to determine if a transaction is
needed as opposed to the rather expensive legacy utxo-based query
approach.

Finally, this also modifies the logic to potentially make more than one
request if needed since it is technically more accurate and avoids any
possibility of error even though the current request sizes are nowhere
near the limits.