mixclient: Avoid nil deref on errored run
When an unactionable error is returned by run(), the logger variable of the
current session is used to log the error, automatically providing additional
context to the log message regarding the session run which errored.  However,
there was still one return statement which returned a nil run state, which
could lead to a panic.