cpuminer: Support discrete miner generate cancel.
Currently, the only way to cancel the discrete mining process once it
has started is to disconnect from the server.

With the aim of ultimately allowing it to be canceled on demand without
disconnecting, this modifies the internal CPU miner discrete mining mode
(aka GenerateNBlocks) to provide the ability to cancel a running
instance from another goroutine by invoking the method with the number
of blocks to generate parameter set to 0.

A running instance that is canceled will return a new exported error
named ErrCancelDiscreteMining.

Note that this only adds the support to the internal CPU miner and does
not yet modify the RPC server to make use of the new functionality.