rpcserver: Discrete mining cancel via generate 0.
This modifies the generate RPC to allow the caller to specify 0 for the
number of blocks parameter in order to cancel the running active
discrete mining instance.

Calling generate 0 when there is not an active instance is a noop and
returns nothing.

Note that that is technically a breaking change to the RPC API since it
currently returns an error when 0 blocks is specified.  However, the
change is deemed acceptable since there are no known callers that rely
on that behavior and not only does the change provide useful
functionality it is also more consistent with the continuous mining
process that is controlled by the setgenerate RPC which accepts 0 as the
number of workers as a way to stop mining.

The existing associated test for the generate RPC is updated for the new
behavior and a new test is added to ensure it is still an error to call
generate with a non-zero number of blocks when there is already an
active discrete mining instance.

Finally, it also updates the associated RPC help strings.