cpuminer: Prioritize num blocks in discrete mining.
The generate RPC was originally built for, and limited to, simnet and
general regression testing in private environments.  As a result, in
order to achieve the primary goal extending the main chain by a specific
number of blocks, it currently assumes there is only ever a single miner
active at a time and just mines how many ever blocks are specified and
returns the hashes of those blocks.

The current approach works well and results in the desired behavior when
there is only a single miner, as is basically always the case with
simnet testing.  However, it produces less than ideal results when there
is any other mining happening concurrently since the final number of
blocks that extend the main chain after it blindly generates the
specified number of blocks might be more or less than the desired amount
depending on what the other miners are doing.

For example, perhaps one node is in the process of executing a generate
10 and while it's mining those 10, another 4 arrive from other sources.
Then the main chain ends up extended by 14 blocks.  Or perhaps several
of those 10 end up as forked blocks due to others showing around the
same time and then the main chain only ends up extended by 7 blocks.

This limitation is particularly relevant now since the assumption that
there is only a single miner is no longer true because the RPC is now
allowed and used on the shared test network which ultimately leads to
additional unnecessary forks and blocks.

Consequently, this reworks the generate RPC semantics (and by extension
the main method in the CPU miner that implements the core logic) to no
longer assume there is only a single active miner.

In particular, it now prioritizes only extending the main chain by the
specified number of blocks regardless of how it happens.  In order to
achieve that goal, the new logic:

* Introduces a new solve context that can be canceled for each mining
  attempt
* Determines the target main chain height based on when the call is
  first invoked and does not terminate until that height is reached
* Only generates as many blocks as are necessary to achieve extending
  the main chain by the specified number of blocks
* Cancels mining attempts when new blocks arrive from other sources
* Prefers building on blocks from other sources that extend the main
  chain
* Discards any locally-mined blocks that would knowingly cause a fork
  with blocks that arrive from other sources
* Discards any locally-mined blocks that would exceed the target height
  due to finishing around the same time as blocks arriving from other
  sources

In keeping with the previous example, it will now only mine 6 blocks
when 4 arrive from other sources.

A summary of the updated semantics from a public-facing standpoint follows:

* It now prioritizes only extending the main chain by the specified
  number of blocks regardless of how it happens
* The returned block hashes are now for the blocks that ultimately
  extended the main chain, regardless of their origin
  * Compare this to the current behavior that returns the hashes of the
    locally-mined blocks regardless of whether or not they ended up on
    the main chain
* There is no observable change for the typical testing scenario of
  only having a single miner active