cpuminer: Terminate discrete mining via blk ntfns.
Currently, the discrete mining behavior, and therefore the generate RPC,
provides slightly inconsistent termination times depending on whether or
not the final block that extends the main chain by the specified target
number of blocks was mined locally or arrives over the network.

Specifically, it terminates immediately when the final block is mined
locally, but waits for a new template when the final block arrives over
the network.  Since new templates are intentionally delayed for a few
seconds to give votes a chance to arrive, the process termination is
also delayed by the same time.

This makes the termination consistent and immediate for all cases by
adding block notifications to the CPU miner when operating in discrete
mining mode and terminating as soon as the target block height is seen
regardless of its origin.