mixclient: Prevent busy looping with canceled local peers
When all local peers are canceled, and there are neither any other remote
peers who have sent KEs for the active session, receiving KEs for the session
from mixpool would immediately error instead of waiting for the context
deadline, causing a busy loop repeatedly retrying the failing session.  This
eventually corrects itself when the PRs are expired and removed from the
mixpool, but until then the mixclient will churn the CPU and output millions
of debug logs.

One way to prevent this is to return an error from the forLocalPeers and
sendLocalPeerMsgs methods if there are no active local peers to perform the
task on.  In run(), the first call to forLocalPeers (to create KEs) will error
and the run() will return instead of failing to complete the pairing followed
by an error forming an alternate session.