blockchain: Sum treasury changes in single db tx.
The current logic that sums up the past treasury balance changes for the
purposes of validation creates a new database transaction for every
block in the range to sum which is inefficient because database
transactions have a fair amount of overhead associated with them.

Thus, this modifies the logic to use a single database transaction for
the entire range of blocks and removes the dbFetchTreasurySingle
function which is no longer used.