blockchain: Remove unneeded assert error.
Blocks are already required to have a treasurybase when the agenda is
active by the time the function is called and so there is no reason to
check the condition again, particularly since any attempt to access the
treasurybase if it didn't actually exist would result in a panic anyway.

This also makes the code more consistent with the coinbase check in the
other branch just under it which also has the exact same semantics in
terms of the existence of the coinbase already having been required to
exist by the time the function is called.