blockchain: Consolidate trsy balance test asserts.
There are several places spread throughout the treasury tests that
duplicate code to assert the treasury balance and associated changes are
the expected values.

Rather than duplicate the checks, this adds a couple of new methods to
the chaingen test harness for that purpose and updates the tests to make
use of them.