chaingen: Add decentralized treasury support.
This adds support for generating blocks that work with the decentralized
treasury to the test chain generator.

It accomplishes this by introducing a new type and method, named
TreasurySemantics and UseTreasurySemantics, respectively, to the chain
generator along with logic to change the behavior accordingly.  The
method takes a parameter of the new type in order to allow the caller
selectively choose which semantics to use.

It also updates the ReplaceWithNVotes munger to work regardless of the
position of the votes within the block as opposed to assuming their
location so that it works equally well with and without a treasurybase
output.

This will ultimately allow all of the treasury tests introduced with
DCP0006 and DCP0007 to be greatly simplified and to significantly reduce
their running time.