blockchain: Use chaingen trsy support in tests.
This updates the treasury tests to minimally make use of the new
decentralized treasury support in chaingen.

Note that this limits the updates to only deal with the areas involving
coinbase and treasurybase transaction changes.

Future commits will further expand the chaingen harness and convert the
tests to simplify them and reduce their running time.