chaingen: Add treasury spend vote support.
This adds support for adding and setting decentralized treasury spend
votes to the test chain generator.

It accomplishes this by introducing two new methods for modifying an
existing stake vote transaction named SetTreasurySpendVote and
AddTreasurySpendVote as well as two new mungers for modifying all of the
stake vote transactions in a block named AddTreasurySpendYesVotes and
AddTreasurySpendNoVotes.

The new methods allow for much more easily adding to or replacing
treasury spend votes as compared to the method the current tests use
which only replace all of them and is rather cumbersome to use and read
since it requires creating slices of structs and repeating a bunch of
parameters at every call site.

This will ultimately allow all of the treasury tests introduced with
DCP0006 and DCP0007 to be further simplified and to significantly reduce
their running time.