dcrec/edwards: Prepare v2.0.4.
This updates the dcrec/edwards module dependencies and serves as a base
for dcrec/edwards/v2.0.4.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/crypto/rand@v1.0.1

The updated indirect dependencies in this commit are as follows:

- golang.org/x/crypto@v0.33.0
- golang.org/x/sys@v0.30.0

The full list of updated direct and indirect dependencies since the
previous dcrec/edwards/v2.0.3 release are the same as above.