chaincfg/chainhash: Prepare v1.0.5.
This updates the chaincfg/chainhash module to use the latest blake256
module dependency and serves as a base for chaincfg/chainhash/v1.0.5.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/crypto/blake256@v1.1.0

The full list of updated direct dependencies since the previous
chaincfg/chainhash/v1.0.4 release are the same as above.