mixclient: Do not cancel mixes before next epoch
This changes the mixing client to continue mixing sessions for however long
they must take to finish or error without requiring this to fit within a
single epoch.  This is a part of a change to avoid a denial of service by
clients who continuously time out at the last possible moment.

Peers matched to mixes are removed from the pending mix requests, potentially
leaving them out of any mixes that begin in the next epoch.