blockchain: Use forced deployments in trsy tests.
This modifies the various treasury tests to make use of the forced
deployment functionality instead of mining a bunch of blocks to first
activate the agenda over and over.

Not only does it simplify the tests and reduce the overall amount of
code quite a bit, it also results in reducing the overall time it takes
to run all of the tests by 3-4x.

This also makes some improvements and does some minor cleanup along the
way such as obtaining the value of a treasurybase based on the
parameters in use from the generated blocks instead of using a global
hard-coded value that prevents other parameters from being used.

There are still several other areas in the tests that could be improved
and most of them could be combined to avoid all of the setup associated
with having them all implemented via separate test funcs, but it's
already quite a nice improvement and large amount of changes without
addressing any of those aspects.