wire: Prepare v1.7.1.
This updates the wire module dependencies and serves as a base for
wire/v1.7.1.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/chaincfg/chainhash@v1.0.5

The updated indirect dependencies in this commit are as follows:

- github.com/decred/dcrd/crypto/blake256@v1.1.0

The full list of updated direct and indirect dependencies since the
previous wire/v1.7.0 release are the same as above.