blockchain/standalone: Prepare v2.2.2
This updates the blockchain/standalone module dependencies and serves as
a base for blockchain/standalone/v2.2.2.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/chaincfg/chainhash@v1.0.5
- github.com/decred/dcrd/wire@v1.7.1

The updated indirect dependencies in this commit are as follows:

- github.com/decred/dcrd/crypto/blake256@v1.1.0

The full list of updated direct and indirect dependencies since the
previous blockchain/standalone/v2.2.1 release are the same as above.

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.