connmgr: Prepare v3.1.3.
This updates the connmgr module dependencies and serves as a base for
connmgr/v3.1.3.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/crypto/rand@v1.0.1
- github.com/decred/dcrd/wire@v1.7.1

The updated indirect dependencies in this commit are as follows:

- github.com/decred/dcrd/chaincfg/chainhash@v1.0.5
- github.com/decred/dcrd/crypto/blake256@v1.1.0
- golang.org/x/crypto@v0.33.0
- golang.org/x/sys@v0.30.0

The full list of updated direct dependencies since the previous
connmgr/v3.1.2 release are the same as above.