gcs: Remove old conditional Go 1.12 code.
Since the module requires a minimum of Go 1.17 now, there is no longer
any reason to keep the conditional compilation of the fast reduction
logic on Go versions prior to 1.12 given it can't build on them anymore
anyway.

With that in mind, this moves the conditionally-compiled implementation
for Go1.13+ into the main file so that it is always compiled as the only
implementation and removes the old conditionally-compiled implementation
for Go versions prior to 1.12.