addrmgr: Add filter parameter to GetAddress
This introduces a breaking change to the address manager by
adding a filter parameter to GetAddress to allow callers to specify
wanted address types.  It also introduces per-bucket statistics to
efficiently track address type counts, which allows tried and new
buckets to be filtered without scanning their contents.  This
allows for identifying whether a bucket contains addresses matching
the network address type filter in constant time rather than
needing to do a linear scan of all addresses within the bucket.
Note that deciding whether a bucket contains any addresses matching
the filter is a distinct operation from selecting an address within
a bucket that is known to contain an address matching the filter.

This prepares for new address types that will be added in the
future.  Filtering will allow callers to prevent address types
that they cannot use or would otherwise be undesirable from being
returned by the address manager.

The following is an overview of the changes:

- Add filter parameter to GetAddress (breaking change)
- Add bucketStats to track address counts per bucket
- Add and update tests