multi: Support TORv3 address type in addrv2
This adds TORv3 support to addrv2 by defining the wire format for TORv3
addresses and updates the type conversion functions between the wire
and addrmgr modules. With this change, TORv3 addresses can be shared
across the peer to peer network for peers with a negotiated protocol
version >= AddrV2Version (12).

The following is an overview of the changes:

wire:
- Add TORv3Address constant (3) to wire.NetAddressType
- Update readNetAddressV2 to handle TORv3Address
- Update writeNetAddressV2 to encode type TORv3Address
- Add tests

server:
- Modify converters to support translating between the wire and addrmgr
  TORv3Address network address type