multi: Enable addrv2 message handling
This adds addrv2 message support to peer and server with
version-aware address exchange, and makes the addrv2 message
active and usable on the peer to peer network for IPv4 and
IPv6 network address types.  With this, the response to a
getaddr message will either be an addr or addrv2 message,
varying based on the negotiated protocol version.

For negotiated protocol version
   <  12, respond to getaddr with addr
   >= 12, respond to getaddr with addrv2

The following is an overview of the changes:

peer:
- Bump peer.MaxProtocolVersion to AddrV2Version (12)
- Add OnAddrV2 callback to MessageListeners
- Add PushAddrV2Msg
- Add MsgAddrV2 handler in inHandler dispatch loop
- Log addrv2 in messageSummary
- Add tests

server:
- Bump maxProtocolVersion to AddrV2Version (12)
- Add isSupportedNetAddressTypeV2 filter for IPv4 and IPv6 only
- Add conversion function wireToAddrmgrNetAddressesV2
- Add pushAddrV2Msg
- Update OnVersion to send addr or addrv2 for negotiated version
- Update OnGetAddr to respond with either addr or addr2
- Add OnAddrV2 handler to process incoming addrv2 messages
- Register OnAddrV2 message listener when constructing peer