wire: Add addrv2 message
This adds the addrv2 message type to allow extensible peer address
exchange. It also defines the protocol version AddrV2Version (12)
for future negotiations, but does not advertise it on the peer to
peer network or use it outside of the wire package.

The structure of the addrv2 message is depicted below for reference.

MsgAddrV2 wire format:
+---------------------------+
| Count (varint, [1..1000]) |
+---------------------------+
| NetAddressV2[N]           |
| (see below)               |
+---------------------------+
| ...                       |
+---------------------------+

The varint Count prefix of the MsgAddrV2 payload describes how many
NetAddressV2 entries exist in the message. This value must indicate
at least 1 entry and no more than 1000 entries to be considered a
valid message.  This differs from the addr message in that it does
not allow zero entries.

NetAddressV2 wire format:
+-----------+----------+--------+----------------+---------+
| Timestamp | Services |  Type  |     Address    |  Port   |
+-----------+----------+--------+----------------+---------+
|  8 bytes  | 8 bytes  | 1 byte | variable bytes | 2 bytes |
+-----------+----------+--------+----------------+---------+

The Type field of the NetAddressV2 payload informs the reader
of both the network type that the address belongs to as well as
allowing the reader to infer how many bytes to read from the
address field.

The addr message is also modified to now fail when encoding/decoding
when the negotiated protocol version >= AddrV2Version.

The following is an overview of the changes:

- Add NetAddressV2 type
  - This includes a constructor NewNetAddressV2 that will be
    needed in a future commit for directly constructing
    v2 wire network addresses with varying types.
- Add MsgAddrV2 type implementing Message interface
- Define AddrV2Version constant (12)
- Reject MsgAddr for negotiated protocol version >= 12
- Add ErrTooFewAddrs error
- Add test coverage