wire: Add addrv2 message
This adds the addrv2 message type to allow extensible peer address
exchange.  The format uses a length-prefixed address field to support
future address types without needing to modify the protocol version.
It also defines the protocol version AddrV2Version (12) for future
negotiations, but does not advertise it on the peer to peer network
or use it outside of the wire package.

The structure of the addrv2 message is depicted below for reference.

MsgAddrV2 wire format:
+---------------------------+
| Count (varint, [1..1000]) |
+---------------------------+
| NetAddressV2[N]           |
| (see below)               |
+---------------------------+
| ...                       |
+---------------------------+

The varint Count prefix of the MsgAddrV2 payload describes how many
NetAddressV2 entries exist in the message. This value must indicate
at least 1 entry and no more than 1000 entries to be considered a
valid message.  This differs from the addr message in that it does
not allow zero entries.

NetAddressV2 wire format:
+-----------+----------+--------+----------+----------------+---------+
| Timestamp | Services |  Type  |  Length  |     Address    |  Port   |
+-----------+----------+--------+----------+----------------+---------+
|  8 bytes  | 8 bytes  | 1 byte |  varint  | [1..512] bytes | 2 bytes |
+-----------+----------+--------+----------+----------------+---------+

The varint Length prefix of the NetAddressV2 payload allows the reader
to gracefully skip unknown address types by reading the specified
number of bytes from the stream, since it is not possible to know the
address type's canoncial size in advance without this field.  This
maintains forward compatibility as new address types are introduced.
The Length value must be greater than or equal to 1 and no more
than 512.

The addr message is also modified to now fail when encoding/decoding
when the negotiated protocol version >= AddrV2Version.

The following is an overview of the changes:

- Add NetAddressV2 type
- Add MsgAddrV2 type implementing Message interface
- Define AddrV2Version constant (12)
- Reject MsgAddr for negotiated protocol version >= 12
- Add ErrTooFewAddrs error
- Add test coverage