peer: Use wire.NetAddressV2 in peer struct
This breaking change modifies the Peer.na field from
wire.NetAddress to wire.NetAddressV2 to accomodate
responses from addrv2.

Since the addrmgr module uses its own NetAddress type,
this adds conversion functions in server.go to translate
between wire.NetAddressV2 and addrmgr.NetAddress.

It also adds an NA function on serverPeer that returns
an addrmgr.NetAddress.  This is done because every
instance of the peer's network address needs to be
converted to an addrmgr.NetAddress anyways,
so shadowing the embedded type removes boilerplate.
Alternatives considered include updating the explicit
conversion function in every usage of NA, couple the
peer module to the addrmgr module and make Peer.na an
addrmgr.NetAddress, or create a new netaddr module to
avoid coupling wire and addrmgr, and adding NA to
serverPeer appeared to be the cleanest approach.

The following is an overview of the changes:

peer:
- Change Peer.na from wire.NetAddress to wire.NetAddressV2
- Add supporting code to convert between NetAddress types

server:
- Update HostToNetAddress callback to return wire.NetAddressV2
- Add NA function on serverPeer to return addrmgr.NetAddress
- Add conversion functions in server.go
- Update OnAddr handler for new types
- Bump peer version to v4 in rpcadaptors.go