server: Enable addrv2 message handling
This enables addrv2 message handling and makes the addrv2 message
active and usable on the peer to peer network for IPv4 and IPv6
network address types.

With this, the response to a getaddr message will either be an
addr or addrv2 message, varying based on the negotiated protocol
version.

For negotiated protocol version
   <  12, respond to getaddr with addr
   >= 12, respond to getaddr with addrv2

The following is an overview of the changes:

- Bump maxProtocolVersion to AddrV2Version (12)
- Add isSupportedNetAddrTypeV1 filter for IPv4 and IPv6 only
- Add conversion function wireToAddrmgrNetAddressesV2
- Add pushAddrV2Msg method
- Update OnVersion to send addr or addrv2 based on negotiated version
- Update OnGetAddr to respond with either addr or addrv2
- Add OnAddrV2 handler to process incoming addrv2 messages
- Register OnAddrV2 message listener when constructing peer