chaincfg: Introduce max treasury spend agenda.
This adds a new definition for the upcoming agenda vote to change the
treasury spend limits to 4% of the total available treasury per
expenditure window as defined in DCP0013.

It does not include any code to make decisions based on the status of
the agenda or bump block versions.  It only adds the definition for the
agenda.