gencerts: Explicit KeyUsage for leaf signing
The "common" use case of using gencerts to generate a self-signed certificate
currently creates a root certificate with the X.509 Key Usage permission to
sign additional leaf certificates.  If this is used to, e.g., generate a
client authentication certificate for dcrwallet, which is added to dcrd's
clients.pem file, this has the surprising security behavior of allowing the
wallet to sign for additional clients that dcrd will happily accept.

To retain the existing, but less common, usage of using gencerts to operate a
CA, a new -S flag must be provided.  This adds back the permission for the
generated certificate to sign its own leaf certificates.