create mixpool observer object at pool creation
both dcrd and wallet use the observer, so there is no point to providing a
separate ctor for it.

this also prevents a data race if you would have called NewObserver
concurrently, as it would have written the mixpool's internal pointer to the
observer (of which only one is tracked) without any synchronization.