mixpool: Add malicious peer greylisting
This implements an observer client that watches for active peers who were
excluded from completed sessions.  This may indicate intentional malicious
behavior (or a buggy peer).  When this behavior is seen too many times, the
peers are marked as (potentially) malicious and their submitted UTXOs become
restricted.

When a new block is mined, and requires votes, no voting notification will be
created if the block contains transactions that double spend UTXOs that were
flagged as malicious by the mixing observer.