chaincfg: Prepare v3.3.0.
This updates the chaincfg module dependencies and serves as a base for
chaincfg/v3.3.0.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/chaincfg/chainhash@v1.0.5
- github.com/decred/dcrd/wire@v1.7.1

The updated indirect dependencies in this commit are as follows:

- github.com/decred/dcrd/crypto/blake256@v1.1.0

The full list of updated direct dependencies since the previous
chaincfg/v3.2.1 release are the same as above.

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.