peer: Prepare v3.2.0.
This updates the peer module dependencies, the copyright year in the
files modified since the previous release, and serves as a base for
peer/v3.2.0.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/chaincfg/chainhash@v1.0.5
- github.com/decred/dcrd/crypto/rand@v1.0.1
- github.com/decred/dcrd/txscript/v4@v4.1.2
- github.com/decred/dcrd/wire@v1.7.1

The updated indirect dependencies in this commit are as follows:

- github.com/decred/dcrd/dcrec/edwards/v2@v2.0.4
- github.com/decred/dcrd/dcrec/secp256k1/v4@v4.4.0
- golang.org/x/crypto@v0.33.0
- golang.org/x/sys@v0.30.0

The full list of updated direct dependencies since the previous
peer/v3.1.3 release are the same as above.

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.