dcrutil: Prepare v4.0.3.
This updates the peer module dependencies and serves as a base for
dcrutil/v4.0.3.

The updated direct dependencies in this commit are as follows:

- github.com/decred/base58@v1.0.6
- github.com/decred/dcrd/chaincfg/chainhash@v1.0.5
- github.com/decred/dcrd/dcrec/edwards/v2@v2.0.4
- github.com/decred/dcrd/dcrec/secp256k1/v4@v4.4.0
- github.com/decred/dcrd/txscript/v4@v4.1.2
- github.com/decred/dcrd/wire@v1.7.1

The updated indirect dependencies in this commit are as follows:

- github.com/decred/dcrd/crypto/blake256@v1.1.0
- github.com/decred/dcrd/crypto/rand@v1.0.1
- golang.org/x/crypto@v0.33.0
- golang.org/x/sys@v0.30.0

The full list of updated direct dependencies since the previous
dcrutil/v4.0.2 release are the same as above.

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.