rpcclient: Prepare v8.1.0.
This updates the rpcclient module dependencies and serves as a base for
rpcclient/v8.1.0.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/chaincfg/chainhash@v1.0.5
- github.com/decred/dcrd/dcrjson/v4@v4.2.0
- github.com/decred/dcrd/dcrutil/v4@v4.0.3
- github.com/decred/dcrd/gcs/v4@v4.1.1
- github.com/decred/dcrd/rpc/jsonrpc/types/v4@v4.3.0
- github.com/decred/dcrd/txscript/v4@v4.1.2
- github.com/decred/dcrd/wire@v1.7.1

The updated indirect dependencies in this commit are as follows:

- github.com/decred/base58@v1.0.6
- github.com/decred/dcrd/blockchain/stake/v5@v5.0.2
- github.com/decred/dcrd/chaincfg/v3@v3.3.0
- github.com/decred/dcrd/crypto/blake256@v1.1.0
- github.com/decred/dcrd/crypto/rand@v1.0.1
- github.com/decred/dcrd/database/v3@v3.0.3
- github.com/decred/dcrd/dcrec/edwards/v2@v2.0.4
- github.com/decred/dcrd/dcrec/secp256k1/v4@v4.4.0
- golang.org/x/crypto@v0.33.0
- golang.org/x/sys@v0.30.0

The full list of updated direct dependencies since the previous
rpcclient/v8.0.1 release are the same as above.

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.