rpc/jsonrpc/types: Prepare v4.4.0.
This updates the rpc/jsonrpc/types module dependencies and serves as a
base for rpc/jsonrpc/types/v4.4.0.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/dcrjson/v4@v4.2.0

The updated indirect dependencies in this commit are as follows:

- github.com/decred/dcrd/chaincfg/chainhash@v1.0.5
- github.com/decred/dcrd/crypto/blake256@v1.1.0

The full list of updated dependencies since the previous
rpc/jsonrpc/types/v4.3.0 release are the same as above.

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.