wire: Pass pointers by interface to writeElement
For at least the past 5 years, due to garbage collector design, Go has not
optimized small non-pointer types to fit inside interface headers, and instead
always uses a pointer to data, if the type is not already a pointer.  This
means that writeElement, which was written before this time to avoid
unnecessary heap allocations, was now causing heap allocations for each
special cased non-pointer type.

This commit rewrites writeElement to only special case pointers to types, and
to pass all elements by pointer.

Co-authored-by: David Hill <dhill@mindcry.org>