wire: Encode header and payload to single buffer
This allows WriteHeaderN to avoid two unnecessary allocations by using the
same buffer to encode both the message header and the encoded message payload.
The payload is appended to zeroed-out header bytes in the buffer, before
calculating the payload checksum and serializing the header to the bytes in
the beginning of the buffer.

This change also results in one less Write to the network connection.