server: Simplify and correct ban score logging.
Currently, the logic that accumulates the persistent and transient ban
scores incorrectly logs debug messages alleging whitelisted peers are
misbehaving due to perfectly normal accumulation of the decaying
transient ban score.

Also, it currently logs the overall ban score whenever it is over the
warning threshold even when the score is not increasing.  There isn't
anything incorrect about that behavior, and the comments even correctly
call out that is what it is intended to do, but it also not very useful.

This modifies the logic to address the aforementioned and takes the
opportunity to simplify the code a bit.

In particular, it now treats whitelisted peers the same as all other
peers in terms of tracking the ban score and warning for misbehavior
whenever the score exceeds the warning threshold.  However, when the ban
score exceeds the ban threshold, whitelisted peers are exempted.

The misbehavior warning messages that are shown once the ban score
exceeds the warning threshold will now include the additional detail
that the peer is whitelisted when that is the case.

Finally, there is no longer any logging when the ban score is not
increasing.